feat #4: Asset Trait added, crate restructuring done to allow this.
