+++ asset-derive/src/variant.rs
