+++ src/lib.rs
@@ -28,9 +28,45 @@
 //!
 //! Please feel free to offer any advice or create a pull request.
 //!
+//!
 //! The original intent of this library was for compile time loading assets
 //! into a binary. This will eventually allow for run-time loading as well,
 //! but as for now that will be a future expansion.
+//!
+//! ## TODOs
+//!
+//! > List of ideas I have at the moment for this project's expansion.
+//! > Please create an issue for a new item to add to this list, using
+//! > `todo` label.
+//!
+//! - [ ] Filename prefix
+//! - [ ] Run-time Loading
+//!     - [ ] Static (Once on init)
+//!     - [ ] Dynamic (Fluid loading)
+//! - [X] ~~Compile-time Loading~~
+//!
+//! ## Structure
+//!
+//! Since `asset-derive` is meant to be a procedural macro crate, while also housing
+//! a trait implementation as well (to be derived), there is a somewhat complex
+//! project structue. This is because of the current annoyance of proc-macro crates
+//! having to be defined completely separate to normal crates.
+//!
+//! The external API shall stay the same fortunately, `asset-derive` will now be stuck
+//! as the trait implementation crate which depends on `asset-derive-macro` which
+//! houses the actual macro implementation. This is unavoidable for the time being, but
+//! I did the best I could to not have the external API change and make it as simple as
+//! can be.
+//!
+//! ### Code Tree
+//!
+//! ```
+//! asset-derive/               <-- Crate to use (trait implementation)
+//!     src/
+//!     examples/               <-- Houses examples using the trait and macro itself.
+//!     asset-derive-macro/     <-- Actual internal derive macro crate. Will be pulled in by main crate.
+//!         src/
+//! ```
 //!
 //! ## Example
 //!
@@ -47,91 +83,16 @@
 //!     FolderDim,
 //! }
 //! ```
-//!
-//! ## TODOs
-//!
-//! > List of ideas I have at the moment for this project's expansion.
-//! > Please create an issue for a new item to add to this list, using
-//! > `todo` label.
-//!
-//! - [ ] Filename prefix
-//! - [ ] Run-time Loading
-//!     - [ ] Static (Once on init)
-//!     - [ ] Dynamic (Fluid loading)
-//! - [X] ~~Compile-time Loading~~
-use proc_macro::TokenStream;
-use quote::quote;
-use syn::{parse_macro_input, DeriveInput, Generics, Ident, Result};
+pub use asset_derive_macro::Asset;
 
-mod attr;
-mod error;
-mod ident;
-mod variant;
-
-use attr::Attributes;
-use error::Error;
-use ident::Identifier;
-use variant::Variants;
-
-/// Assets type.
-///
-/// Houses the entire parsed macro contents.
-///
-/// * `name`: Main enum identifier.
-/// * `generics`: Any generic information.
-/// * `attrs`: Base level attributes.
-/// * `variants`: All variants.
-#[derive(Debug)]
-struct Assets<'a> {
-    name: &'a Ident,
-    generics: &'a Generics,
-    attrs: Attributes,
-    variants: Variants,
-}
-
-impl<'a> Assets<'a> {
-    /// Creates an `Assets` type from parsed macro input.
+/// Trait to be derived by `asset-derive` macro.
+pub trait Asset {
+    /// Method responsible for fetching requested resource.
     ///
-    /// * `input`: Parsed macro input.
-    pub fn from(input: &'a DeriveInput) -> Result<Self> {
-        Ok(Self {
-            name: &input.ident,
-            generics: &input.generics,
-            attrs: Attributes::from(&input.attrs)?,
-            variants: Variants::from(input)?,
-        })
-    }
-
-    /// Builds the entire macro output.
-    pub fn build(self) -> TokenStream {
-        let (impl_generics, ty_generics, where_clause) = self.generics.split_for_impl();
-        let name = self.name;
-        let getters = self.variants.build_getters(&self.attrs);
-        let arms = self.variants.build_arms();
-        quote! {
-            impl #impl_generics #name #ty_generics #where_clause {
-                #(#getters)*
-                pub fn fetch(&self) -> Vec<u8> {
-                    match self {
-                        #(#arms),*
-                    }.to_vec()
-                }
-            }
-        }
-        .into()
-    }
-}
-
-/// Entry function of the derive macro.
-#[proc_macro_derive(Asset, attributes(asset))]
-pub fn derive_asset(input: TokenStream) -> TokenStream {
-    impl_asset(&parse_macro_input!(input as DeriveInput))
-        .unwrap_or_else(|err| err.to_compile_error().into())
-}
-
-/// Lower implementation of `Asset`.
-///
-/// * `input`: Parsed macro input.
-fn impl_asset(input: &DeriveInput) -> Result<TokenStream> {
-    Ok(Assets::from(input)?.build())
+    /// ## Example
+    ///
+    /// ```
+    /// let data = Icon::Select.fetch();
+    /// ```
+    fn fetch(&self) -> Vec<u8>;
 }
